feature: 原子操作能力支持滚动执行 #446
agent_task 新增 actual_execute_count字段用于表示实际重试的次数

+++ src/backend/job-execute/service-job-execute/src/main/java/com/tencent/bk/job/execute/dao/impl/StepInstanceRollingTaskDAOImpl.java
@@ -71,7 +71,7 @@ public class StepInstanceRollingTaskDAOImpl implements StepInstanceRollingTaskDA
         Record record = CTX.select(ALL_FIELDS)
             .from(TABLE)
             .where(TABLE.STEP_INSTANCE_ID.eq(stepInstanceId))
-            .and(TABLE.EXECUTE_COUNT.eq(JooqDataTypeUtil.toByte(executeCount)))
+            .and(TABLE.EXECUTE_COUNT.eq(JooqDataTypeUtil.toShort(executeCount)))
             .and(TABLE.BATCH.eq(JooqDataTypeUtil.toShort(batch)))
             .fetchOne();
         return extract(record);
@@ -101,7 +101,7 @@ public class StepInstanceRollingTaskDAOImpl implements StepInstanceRollingTaskDA
             .from(TABLE)
             .where(TABLE.STEP_INSTANCE_ID.eq(stepInstanceId));
         if (executeCount != null) {
-            selectConditionStep.and(TABLE.EXECUTE_COUNT.eq(executeCount.byteValue()));
+            selectConditionStep.and(TABLE.EXECUTE_COUNT.eq(executeCount.shortValue()));
         }
         if (batch != null && batch > 0) {
             selectConditionStep.and(TABLE.BATCH.eq(batch.shortValue()));
@@ -129,7 +129,7 @@ public class StepInstanceRollingTaskDAOImpl implements StepInstanceRollingTaskDA
             TABLE.TOTAL_TIME)
             .values(
                 rollingTask.getStepInstanceId(),
-                JooqDataTypeUtil.toByte(rollingTask.getExecuteCount()),
+                JooqDataTypeUtil.toShort(rollingTask.getExecuteCount()),
                 JooqDataTypeUtil.toShort(rollingTask.getBatch()),
                 JooqDataTypeUtil.toByte(rollingTask.getStatus()),
                 rollingTask.getStartTime(),
@@ -179,7 +179,7 @@ public class StepInstanceRollingTaskDAOImpl implements StepInstanceRollingTaskDA
             return;
         }
         updateSetMoreStep.where(TABLE.STEP_INSTANCE_ID.eq(stepInstanceId))
-            .and(TABLE.EXECUTE_COUNT.eq(JooqDataTypeUtil.toByte(executeCount)))
+            .and(TABLE.EXECUTE_COUNT.eq(JooqDataTypeUtil.toByte(executeCount).shortValue()))
             .and(TABLE.BATCH.eq(JooqDataTypeUtil.toShort(batch)))
             .execute();
